Module kms::api

Api definitions for KMS/Vault drivers

To establish a connection between the client and the KMS/Vault server, two kinds of parameters are required:

Public parameters: like Region Id of the KMS/Vault, Instance Id of the KMS, etc. They are not confidential and can be seen by anyone. ProviderSettings is a json object. It is to include all the public parameters. The hashmap style makes it flexible for different kinds of KMSes/Vaults. There should be a function that takes a ProviderSettings as a input parameter and returns a client instance.

Private parameters: like the credential to access (e.g. access key). These parameters should be captured inside the logic of new() rather than the input parameter. it is strongly recommended that private parameters be read from the encrypted filesystem, e.g. /run/* which is in TEE’s encrypted memory.

APIs

Decrypter: KMS’s decrypt API.

Encrypter: KMS’s encrypt API.

Getter: Vault’s get secret API.

Setter: Vault’s set secret API.

The rationality to distinguish these four different traits:

Decrypter and Getter are used in-guest, while Encrypter and Setter are used userside. They do not need to be implemented by a same object.

Traits

Decrypter

Encrypter

Getter

Setter

Type Aliases

Annotations

Annotations is extra information of this encryption/decryption.

ProviderSettings

ProviderSettings are extra information to create a client