Module kms::plugins

Modules

aliyun

This is a Aliyun KMS implementation.

kbs

Abstraction for KBCs as a KMS plugin.

Enums

DecryptorProvider

VaultProvider

Functions

new_decryptor

Create a new Decrypter by given provider name and ProviderSettings

new_getter

Create a new Getter by given provider name and ProviderSettings